backups/2025-07-19_22-44-19_production_ready/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.1.2
black==24.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coverage==7.8.2
cryptography==42.0.8
deprecation==2.1.0
django-encrypted-model-fields==0.6.3
flake8==7.0.0
flasgger==0.9.7.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
gotrue==2.11.4
greenlet==3.2.2
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mccabe==0.7.0
mistune==3.1.3
multidict==6.4.4
mypy_extensions==1.1.0
numpy==1.26.4
packaging==25.0
pandas==2.2.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
pycodestyle==2.11.1
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
referencing==0.36.2
requests==2.31.0
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.20.0
